+++ docs/data/date-pickers/date-range-field/BasicDateRangeField.tsx
@@ -6,20 +6,19 @@ import { MultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDate
 import { SingleInputDateRangeField } from '@mui/x-date-pickers-pro/SingleInputDateRangeField';
 
 export default function BasicDateRangeField() {
-  const [value, setValue] = React.useState<any>([null, null]);
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <DemoContainer
         components={['MultiInputDateRangeField', 'SingleInputDateRangeField']}
       >
         <MultiInputDateRangeField
-          value={value}
-          onChange={(newValue) => setValue(newValue)}
+          slotProps={{
+            textField: ({ position }) => ({
+              label: position === 'start' ? 'Departure' : 'Return',
+            }),
+          }}
         />
-        <SingleInputDateRangeField
-          value={value}
-          onChange={(newValue) => setValue(newValue)}
-        />
+        <SingleInputDateRangeField label="Departure - Return" />
       </DemoContainer>
     </LocalizationProvider>
   );