+++ packages/grid/x-data-grid/src/components/panel/filterPanel/GridFilterPanel.tsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { SxProps, Theme } from '@mui/material/styles';
 import { GridFilterItem, GridLogicOperator } from '../../../models/gridFilterItem';
 import { useGridApiContext } from '../../../hooks/utils/useGridApiContext';
-import { GridAddIcon } from '../../icons';
 import { GridPanelContent } from '../GridPanelContent';
 import { GridPanelFooter } from '../GridPanelFooter';
 import { GridPanelWrapper } from '../GridPanelWrapper';
@@ -44,6 +43,17 @@ export interface GridFilterPanelProps
     | 'valueInputProps'
   >;
 
+  /*
+   * If `true`, the `Add filter` button will not be displayed.
+   * @default false
+   */
+  disableAddFilterButton?: boolean;
+  /*
+   * If `true`, the `Remove all` button will be disabled
+   * @default false
+   */
+  disableRemoveAllButton?: boolean;
+
   /**
    * @ignore - do not document.
    */
@@ -70,6 +80,8 @@ const GridFilterPanel = React.forwardRef<HTMLDivElement, GridFilterPanelProps>(
       filterFormProps,
       getColumnForNewFilter,
       children,
+      disableAddFilterButton = false,
+      disableRemoveAllButton = false,
       ...other
     } = props;
 
@@ -165,6 +177,14 @@ const GridFilterPanel = React.forwardRef<HTMLDivElement, GridFilterPanelProps>(
       [apiRef, items.length],
     );
 
+    const handleRemoveAll = () => {
+      if (items.length === 1 && items[0].value === undefined) {
+        apiRef.current.deleteFilterItem(items[0]);
+        apiRef.current.hideFilterPanel();
+      }
+      apiRef.current.setFilterModel({ ...filterModel, items: [] });
+    };
+
     React.useEffect(() => {
       if (
         logicOperators.length > 0 &&
@@ -202,17 +222,33 @@ const GridFilterPanel = React.forwardRef<HTMLDivElement, GridFilterPanelProps>(
             />
           ))}
         </GridPanelContent>
-        {!rootProps.disableMultipleColumnsFiltering && (
+        {!rootProps.disableMultipleColumnsFiltering &&
+        !disableAddFilterButton &&
+        !disableRemoveAllButton ? (
           <GridPanelFooter>
-            <rootProps.components.BaseButton
-              onClick={addNewFilter}
-              startIcon={<GridAddIcon />}
-              {...rootProps.componentsProps?.baseButton}
-            >
-              {apiRef.current.getLocaleText('filterPanelAddFilter')}
-            </rootProps.components.BaseButton>
+            {!disableAddFilterButton ? (
+              <rootProps.slots.baseButton
+                onClick={addNewFilter}
+                startIcon={<rootProps.slots.filterPanelAddIcon />}
+                {...rootProps.slotProps?.baseButton}
+              >
+                {apiRef.current.getLocaleText('filterPanelAddFilter')}
+              </rootProps.slots.baseButton>
+            ) : (
+              <span />
+            )}
+
+            {!disableRemoveAllButton ? (
+              <rootProps.slots.baseButton
+                onClick={handleRemoveAll}
+                startIcon={<rootProps.slots.filterPanelRemoveAllIcon />}
+                {...rootProps.slotProps?.baseButton}
+              >
+                {apiRef.current.getLocaleText('filterPanelRemoveAll')}
+              </rootProps.slots.baseButton>
+            ) : null}
           </GridPanelFooter>
-        )}
+        ) : null}
       </GridPanelWrapper>
     );
   },
@@ -232,6 +268,8 @@ GridFilterPanel.propTypes = {
    * If not specified, the order is derived from the `columns` prop.
    */
   columnsSort: PropTypes.oneOf(['asc', 'desc']),
+  disableAddFilterButton: PropTypes.bool,
+  disableRemoveAllButton: PropTypes.bool,
   /**
    * Props passed to each filter form.
    */