+++ packages/grid/x-data-grid/src/constants/localeTextConstants.ts
@@ -45,6 +45,7 @@ export const GRID_DEFAULT_LOCALE_TEXT: GridLocaleText = {
 
   // Filter panel text
   filterPanelAddFilter: 'Add filter',
+  filterPanelRemoveAll: 'Remove all',
   filterPanelDeleteIconLabel: 'Delete',
   filterPanelLogicOperator: 'Logic operator',
   filterPanelOperator: 'Operator',