+++ packages/grid/x-data-grid/src/internals/index.ts
@@ -31,6 +31,7 @@ export {
   useGridColumnGrouping,
   columnGroupsStateInitializer,
 } from '../hooks/features/columnGrouping/useGridColumnGrouping';
+export type { GridColumnGroupLookup } from '../hooks/features/columnGrouping/gridColumnGroupsInterfaces';
 export type {
   GridColumnRawLookup,
   GridColumnsRawState,
@@ -118,3 +119,5 @@ export { useGridPrivateApiContext } from '../hooks/utils/useGridPrivateApiContex
 
 export type { GridApiCommunity } from '../models/api/gridApiCommunity';
 export type { GridApiCaches } from '../models/gridApiCaches';
+
+export * from './utils';