+++ packages/x-date-pickers-pro/src/index.ts
@@ -29,5 +29,7 @@ export * from './StaticDateRangePicker';
 export * from './dateRangeViewRenderers';
 
 export type { DateRangeValidationError } from './internal/hooks/validation/useDateRangeValidation';
+export type { DateTimeRangeValidationError } from './internal/hooks/validation/useDateTimeRangeValidation';
+export type { TimeRangeValidationError } from './internal/hooks/validation/useTimeRangeValidation';
 export type { DateRange } from './internal/models/range';
 export type { UseDateRangeFieldProps } from './internal/models/dateRange';