+++ packages/x-date-pickers/src/tests/fieldKeyboardInteraction.test.tsx
@@ -1,18 +1,24 @@
 import * as React from 'react';
-import { screen, userEvent, act } from '@mui/monorepo/test/utils';
-import { createPickerRenderer, expectInputValue } from 'test/utils/pickers-utils';
+import moment from 'moment/moment';
+import jMoment from 'moment-jalaali';
+import { screen, userEvent } from '@mui/monorepo/test/utils';
+import {
+  buildFieldInteractions,
+  createPickerRenderer,
+  expectInputValue,
+} from 'test/utils/pickers-utils';
 import { DateTimeField } from '@mui/x-date-pickers/DateTimeField/DateTimeField';
-import { MuiDateSectionName, MuiPickersAdapter } from '../internals/models/muiPickersAdapter';
+import { FieldSectionType, MuiPickersAdapter } from '../internals/models/muiPickersAdapter';
 
 const testDate = new Date(2018, 4, 15, 9, 35, 10);
 
-function updatedDate<TDate>(
+function updateDate<TDate>(
   date: TDate,
   adapter: MuiPickersAdapter<TDate>,
-  sectionName: MuiDateSectionName,
+  sectionType: FieldSectionType,
   diff: number,
 ) {
-  switch (sectionName) {
+  switch (sectionType) {
     case 'year':
       return adapter.addYears(date, diff);
     case 'month':
@@ -39,7 +45,7 @@ const adapterToTest = [
   'dayjs',
   'moment',
   'date-fns-jalali',
-  'moment-hijri',
+  // 'moment-hijri',
   'moment-jalaali',
 ] as const;
 
@@ -50,13 +56,21 @@ adapterToTest.forEach((adapterName) => {
       adapterName,
     });
 
-    const clickOnInput = (input: HTMLInputElement, cursorPosition: number) => {
-      act(() => {
-        input.focus();
-        input.setSelectionRange(cursorPosition, cursorPosition);
-        clock.runToLast();
-      });
-    };
+    before(() => {
+      if (adapterName === 'moment-jalaali') {
+        jMoment.loadPersian();
+      } else if (adapterName === 'moment') {
+        moment.locale('en');
+      }
+    });
+
+    after(() => {
+      if (adapterName === 'moment-jalaali') {
+        moment.locale('en');
+      }
+    });
+
+    const { clickOnInput } = buildFieldInteractions({ clock, render, Component: DateTimeField });
 
     const testKeyPress = <TDate extends unknown>({
       key,
@@ -75,14 +89,16 @@ adapterToTest.forEach((adapterName) => {
       const input = screen.getByRole('textbox') as HTMLInputElement;
       clickOnInput(input, cursorPosition);
       userEvent.keyPress(input, { key });
+
       expectInputValue(input, adapter.formatByString(expectedValue, format));
     };
 
-    const testKeyboardInteraction = (formatToken, sectionData) => {
-      const sectionName = typeof sectionData === 'object' ? sectionData.sectionName : sectionData;
-      it(`should increase "${sectionName}" when pressing ArrowUp on "${formatToken}" token`, () => {
+    const testKeyboardInteraction = (formatToken, sectionConfig) => {
+      const sectionType =
+        typeof sectionConfig === 'object' ? sectionConfig.sectionType : sectionConfig;
+      it(`should increase "${sectionType}" when pressing ArrowUp on "${formatToken}" token`, () => {
         const initialValue = adapter.date(testDate);
-        const expectedValue = updatedDate(initialValue, adapter, sectionName, 1);
+        const expectedValue = updateDate(initialValue, adapter, sectionType, 1);
 
         testKeyPress({
           key: 'ArrowUp',
@@ -92,9 +108,9 @@ adapterToTest.forEach((adapterName) => {
         });
       });
 
-      it(`should decrease "${sectionName}" when pressing ArrowDown on "${formatToken}" token`, () => {
+      it(`should decrease "${sectionType}" when pressing ArrowDown on "${formatToken}" token`, () => {
         const initialValue = adapter.date(testDate);
-        const expectedValue = updatedDate(initialValue, adapter, sectionName, -1);
+        const expectedValue = updateDate(initialValue, adapter, sectionType, -1);
 
         testKeyPress({
           key: 'ArrowDown',
@@ -105,8 +121,8 @@ adapterToTest.forEach((adapterName) => {
       });
     };
 
-    Object.entries(adapter.formatTokenMap).forEach(([formatToken, sectionName]) =>
-      testKeyboardInteraction(formatToken, sectionName),
+    Object.entries(adapter.formatTokenMap).forEach(([formatToken, sectionConfig]) =>
+      testKeyboardInteraction(formatToken, sectionConfig),
     );
   });
 });